Updates README Links to Use the Correct Line #s
There were two references to `PinpointKitDelegate` pointing to the middle of a function, so I’ve updated those to point to the declaration of the protocol. If they should be on the `PinpointKitDelegate` property in `PinpointKit`, we could do that instead, but that wouldn't match the links to `Configuration`, which currently link to the `Configuration` file. (no line #s).

The only other line number referenced in the realm is for [`defaultPinpointKit`](https://github.com/Lickability/PinpointKit/blob/master/PinpointKit/PinpointKit/Sources/PinpointKit.swift#L17) which is currently correct.

I also removed the word "a" from the sentence "...you can specify both a `Configuration` and `PinpointKitDelegate` instances..." because "a" is singular and "instances" is plural, causing this to read a bit awkwardly.